Moving the args to helper
